+++ packages/web-core/src/factories/event.ts
@@ -1,6 +1,7 @@
 import { v4 } from 'uuid';
 
 import {
+  Actor,
   ActorType,
   Agent,
   AudioSessionStartPayloadMicrophoneMode,
@@ -27,7 +28,6 @@ import { Character } from '../entities/character.entity';
 export interface SendCancelResponsePacketParams {
   interactionId?: string;
   utteranceId?: string[];
-  character: Character;
 }
 
 export class EventFactory {
@@ -108,7 +108,9 @@ export class EventFactory {
     return {
       ...this.baseProtoPacket({ correlationId: true, conversationId }),
       ...(character && {
-        routing: this.routing({ character }),
+        routing: this.routing({
+          target: { name: character.id, type: ActorType.AGENT },
+        }),
       }),
       custom: {
         name,
@@ -130,7 +132,9 @@ export class EventFactory {
           utteranceId: params.utteranceId,
         },
       },
-      routing: this.routing({ character: params.character }),
+      routing: this.routing({
+        target: { type: ActorType.WORLD },
+      }),
     };
   }
 
@@ -307,12 +311,10 @@ export class EventFactory {
     };
   }
 
-  private routing(props?: { character: Character }): Routing {
+  private routing(props?: { target: Actor }): Routing {
     return {
       source: { type: ActorType.PLAYER },
-      ...(props?.character && {
-        target: { type: ActorType.AGENT, name: props.character.id },
-      }),
+      ...(props?.target && { target: props.target }),
     };
   }